requirements.txt
# ── Core (always required) ─────────────────────────────────────────────────
av>=12.0
Pillow>=10.0
numpy>=1.24
pydantic>=2.0
typer>=0.12
fastapi>=0.110
uvicorn>=0.29
aiosqlite>=0.19
pyyaml>=6.0
rich>=13.0

# ── Local model backends ───────────────────────────────────────────────────
ollama>=0.2                     # Ollama backend (pip install localvisionai[ollama])

# ── Cloud model backends ───────────────────────────────────────────────────
openai>=1.30                    # OpenAI + LM Studio backends
anthropic>=0.28                 # Anthropic / Claude backend
google-generativeai>=0.8        # Google Gemini backend

# ── HuggingFace Transformers backend ──────────────────────────────────────
transformers>=4.40
# accelerate>=0.28
# bitsandbytes>=0.43
# sentencepiece>=0.2
torch>=2.2

# ── Other optional local backends (uncomment as needed) ───────────────────
# llama-cpp-python>=0.2.70      # llama.cpp backend
# mlx>=0.12                     # MLX backend (Apple Silicon only)
# mlx-vlm>=0.1
# vllm>=0.4                     # vLLM backend

# ── Optional input sources ─────────────────────────────────────────────────
opencv-python>=4.8            # Webcam + scene detection
scenedetect>=0.6              # Scene-change sampler
mss>=9.0                      # Screen capture source
yt-dlp>=2024.1                # YouTube / URL source

# ── Optional extras ───────────────────────────────────────────────────────
pytesseract>=0.3              # OCR support
openai-whisper>=20231117      # Audio transcription
chromadb>=0.5                 # RAG / vector search
# sentence-transformers>=2.7
# gradio>=4.0                   # Gradio UI
aiohttp>=3.9                  # Web utilities

# ── Development ────────────────────────────────────────────────────────────
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
ruff>=0.4
mypy>=1.10
httpx>=0.27
